support loading profile in web - embedder api to pass profile and contents - initialise the profile
