+++ src/vs/workbench/services/userDataProfile/browser/keybindingsResource.ts
@@ -6,7 +6,7 @@
 import { VSBuffer } from 'vs/base/common/buffer';
 import { FileOperationError, FileOperationResult, IFileService } from 'vs/platform/files/common/files';
 import { ILogService } from 'vs/platform/log/common/log';
-import { IProfileResource, IProfileResourceChildTreeItem, IProfileResourceTreeItem, ProfileResourceType } from 'vs/workbench/services/userDataProfile/common/userDataProfile';
+import { IProfileResource, IProfileResourceChildTreeItem, IProfileResourceInitializer, IProfileResourceTreeItem, IUserDataProfileService, ProfileResourceType } from 'vs/workbench/services/userDataProfile/common/userDataProfile';
 import { platform, Platform } from 'vs/base/common/platform';
 import { ITreeItemCheckboxState, TreeItemCollapsibleState } from 'vs/workbench/common/views';
 import { IUserDataProfile } from 'vs/platform/userDataProfile/common/userDataProfile';
@@ -17,6 +17,25 @@ import { localize } from 'vs/nls';
 interface IKeybindingsResourceContent {
 	platform: Platform;
 	keybindings: string | null;
+}
+
+export class KeybindingsResourceInitializer implements IProfileResourceInitializer {
+
+	constructor(
+		@IUserDataProfileService private readonly userDataProfileService: IUserDataProfileService,
+		@IFileService private readonly fileService: IFileService,
+		@ILogService private readonly logService: ILogService,
+	) {
+	}
+
+	async initialize(content: string): Promise<void> {
+		const keybindingsContent: IKeybindingsResourceContent = JSON.parse(content);
+		if (keybindingsContent.keybindings === null) {
+			this.logService.info(`Initializing Profile: No keybindings to apply...`);
+			return;
+		}
+		await this.fileService.writeFile(this.userDataProfileService.currentProfile.keybindingsResource, VSBuffer.fromString(keybindingsContent.keybindings));
+	}
 }
 
 export class KeybindingsResource implements IProfileResource {
